need to nuke EF and setup new connection

+++ algo-02/algo-02/LogicLayer/Modelinterface.cs
@@ -18,6 +18,7 @@ namespace algo_02.LogicLayer
             using (var context = new DatabaseContext())
             {
                 //select table by table in order of dependant to non dependant and nuke all data
+
                 List<SYMBOL_HISTORY> symbolHistoryCollection = (from x in context.SYMBOL_HISTORY
                                                                 select x).ToList();
                 List<WALLET_HISTORY> walletHistoryCollection = (from x in context.WALLET_HISTORY
@@ -31,31 +32,47 @@ namespace algo_02.LogicLayer
                 List<WatchList> watchListCollection = (from x in context.WatchLists
                                                        select x).ToList();
 
-                foreach (var item in symbolHistoryCollection)
+                if (portfolioCollection != null)
                 {
-                    context.SYMBOL_HISTORY.Remove(item);
+
+                    foreach (var item in symbolHistoryCollection)
+                    {
+                        context.SYMBOL_HISTORY.Remove(item);
+                    }
+                    foreach (var item in walletHistoryCollection)
+                    {
+                        context.WALLET_HISTORY.Remove(item);
+                    }
+                    foreach (var item in walletCollection)
+                    {
+                        context.Wallets.Remove(item);
+                    }
+                    foreach (var item in portfolioCollection)
+                    {
+                        context.Portfolios.Remove(item);
+
+                    }
+                    foreach (var item in stockCollection)
+                    {
+                        context.Stock_Item.Remove(item);
+                    }
+                    foreach (var item in watchListCollection)
+                    {
+                        context.WatchLists.Remove(item);
+                    }
+                    try
+                    {
+                        context.SaveChanges();
+                    }
+                    catch (Exception e)
+                    {
+
+                        Console.WriteLine("init nuke " + e.InnerException);
+                    }
                 }
-                foreach (var item in walletHistoryCollection)
-                {
-                    context.WALLET_HISTORY.Remove(item);
-                }
-                foreach (var item in walletCollection)
-                {
-                    context.Wallets.Remove(item);
-                }
-                foreach (var item in portfolioCollection)
-                {
-                    context.Portfolios.Remove(item);
-                }
-                foreach (var item in stockCollection)
-                {
-                    context.Stock_Item.Remove(item);
-                }
-                foreach (var item in watchListCollection)
-                {
-                    context.WatchLists.Remove(item);
-                }
-                context.SaveChanges();
+
+
+
 
             }
         }
